+++ R/makeGroupMembers.R
@@ -20,36 +20,10 @@ makeGroupMembers <- function(numGp, currentGroup, candidates, ped, harem,
                              minAge) {
   groupMembers <- list()
   if (harem) {
-    if (length(getPotentialSires(currentGroup, minAge, ped)) > 1)
-      stop(paste0("User selected to form harems with more than one male, ",
-                  "There are ",
-                  length(getPotentialSires(currentGroup, minAge, ped)), " at ",
-                  " least ", minAge, " years old in the current group."))
-    if (length(currentGroup) > 0 & numGp > 1)
-      stop(paste0("User cannot have more than one breeding group with ",
-                  "animals in a current group"))
-    if (length(getPotentialSires(candidates, minAge, ped)) < numGp &
-        length(getPotentialSires(currentGroup, minAge, ped)) == 0)
-      stop(paste0("User selected to form harems in ", numGp, " groups with ",
-                  "only ", length(getPotentialSires(currentGroup, minAge, ped)),
-                  " males at least ",
-                  minAge, " years old in the list of candidates."))
-
-    if (length(getPotentialSires(currentGroup, minAge, ped)) == 0) {
-      ped <- ped[!is.na(ped$birth), ]
-      sires <- sample(getPotentialSires(candidates, minAge, ped), numGp,
-                      replace = FALSE)
-      for (i in 1:numGp) {
-        groupMembers[[i]] <- sires[i]
-      }
-    }
-    if (length(currentGroup) > 0) {
-      if (length(getPotentialSires(currentGroup, minAge, ped)) > 0) {
-        groupMembers[[1]] <- currentGroup
-      } else {
-        groupMembers[[1]] <- c(groupMembers[[1]], currentGroup)
-      }
-    }
+    ## Since harems only have a single male, they are inserted during
+    ## initialization.
+    groupMembers <- initializeHaremGroups(numGp, currentGroup, candidates,
+                                          ped, minAge)
   } else {
     for (i in 1:numGp) {
       if (length(currentGroup) == 0) {