+++ NEWS.md
@@ -3,8 +3,11 @@ NEWS
 R. Mark Sharp
 12/30/2018
 
-# nprcmanager 0.5.01 (20181230)
+# nprcmanager 0.5.02 (20190103)
 
+  - Added ability to specify sex ratio in increments of 0.5
+    (Female/Male) from 0.5 to 10 in increments of 0.5. \# nprcmanager
+    0.5.01 (20181230)
   - Correction of some bugs in harem creation and provided additional
     unit tests for harem creation to prevent regression. \# nprcmanager
     0.5.00 (20181228)